LPS-193631 Rename DataRecordModelResourcePermission to DataRecordPermissionUtil and apply to usages
